app/backend/requirements.txt
fastapi == 0.89.1

uvicorn == 0.20.0

numpy == 1.24.2
scipy == 1.10.0
pandas==1.5.3

sktime==0.26.0
scikit-learn==1.4.0

umap-learn==0.5.5

captum==0.6.0

torch==1.12.1
torchvision==0.13.1

dill==0.3.8
